refactor: 게시물 리스트 하위컴포넌트 분리
