+++ main-camp/src/app/boards/page.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import BoardsList from "@/components/boards-list"
+import BoardsList from "@/components/boards-list/list"
 
 const BoardsListPage = () => {
     return (
