+++ main-camp/src/components/boards-list/list/index.tsx
