+++ main-camp/src/components/boards-list/list/type.ts
